pyproject.toml
[tool.poetry]
name = "chyme"
version = "0.1.0"
description = ""
authors = ["petelouise <programmetress@petelouise.beauty>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
vcrpy = "^6.0.1"
pytest-recording = "^0.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
